+++ src/mympd_api/home.h
@@ -16,12 +16,21 @@
 
 #include <stdbool.h>
 
+/**
+ * Home widget types
+ */
+enum home_type {
+    HOME_ICON,
+    HOME_WIDGET_SCRIPT,
+    HOME_WIDGET_IFRAME
+};
+
 bool mympd_api_home_icon_move(struct t_list *home_list, unsigned from, unsigned to);
 bool mympd_api_home_icon_delete(struct t_list *home_list, unsigned pos);
 bool mympd_api_home_icon_save(struct t_list *home_list, bool replace, unsigned oldpos,
     sds name, sds ligature, sds bgcolor, sds color, sds image, sds cmd, struct t_list *option_list);
 bool mympd_api_home_widget_save(struct t_list *home_list, bool replace, unsigned oldpos,
-    sds name, unsigned refresh, sds size, sds script, struct t_list *arguments);
+    enum home_type type, sds name, unsigned refresh, sds size, sds ref, struct t_list *arguments);
 sds mympd_api_home_icon_list(struct t_list *home_list, sds buffer, unsigned request_id);
 sds mympd_api_home_icon_get(struct t_list *home_list, sds buffer, unsigned request_id, unsigned pos);
 bool mympd_api_home_file_read(struct t_list *home_list, sds workdir);